Added Antlr compilation to test targets.
Test target now builds the Antlr runtime (so it can be the first thing run when the repo is cloned) and runs Antlr (so the parser gets updated when we change the grammar and want to test it). In practice, these scripts should return immediately almost every time, so testing time should remain largely unaffected.